requirements.txt
mysql-connector-python==8.3.0
tabulate==0.9.0